refactored api to a static server to deploy to heroku instaed of building custom runtime for AWS lambda
