+++ .gitignore
@@ -1,3 +1,4 @@
+# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
 /node_modules
@@ -21,16 +22,3 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-.vercel
-lambda/bundle
-lambda/dist
-
-
-# Local Netlify folder
-.netlify
-
-graphql.schema.json
-
-codegen.ts
-
-src/seed
