--- codegen.yml
@@ -1,10 +1,0 @@
-overwrite: true
-schema: 'http://localhost:4000/graphql'
-generates:
-  ./src/generated/graphql.ts:
-    plugins:
-      - 'typescript'
-      - 'typescript-resolvers'
-  ./graphql.schema.json:
-    plugins:
-      - 'introspection'
