--- lambda/tsconfig.json
@@ -1,17 +1,0 @@
-{
-	"compilerOptions": {
-		"sourceMap": true,
-		"noImplicitAny": true,
-		"module": "commonjs",
-		"target": "es6",
-		"experimentalDecorators": true,
-		"allowSyntheticDefaultImports": true,
-		"moduleResolution": "node",
-		"skipLibCheck": true,
-		"esModuleInterop": true,
-		"outDir": "./dist",
-		"resolveJsonModule": true
-	},
-	"include": ["./*.ts", "./**/*.ts", "./**/*.js"],
-	"exclude": ["./generated/*.js"]
-}
