+++ package.json
@@ -1,48 +1,52 @@
 {
 	"name": "south-shore-mech-svcs-consumer-api",
-	"version": "1.0.0",
-	"description": "Lambda function for South Shore Mechanical Services Consumer Website",
-	"main": "index.ts",
+	"version": "2.0.0",
+	"description": "",
+	"main": "server.ts",
 	"scripts": {
-		"start": "nodemon",
-		"offline": "npm run build:lambda && serverless offline",
-		"build:lambda": "npm run bundle && tsc -p lambda/tsconfig.json",
-		"codegen": "graphql-codegen --config codegen.yml",
-		"bundle": "ncp ./src ./lambda/bundle"
+		"start": "ts-node server.ts",
+		"dev": "nodemon server.ts",
+		"watch": "nodemon server.ts",
+		"build": "npm run minify && webpack",
+		"seed": "node seeders/seed.js",
+		"codegen": "graphql-codegen --config codegen.ts"
 	},
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
 	"dependencies": {
-		"@apollo/server": "^4.10.4",
-		"@as-integrations/aws-lambda": "^3.1.0",
 		"@aws-sdk/client-s3": "^3.583.0",
 		"@aws-sdk/credential-providers": "^3.583.0",
 		"@aws-sdk/s3-request-presigner": "^3.583.0",
-		"apollo-server": "^3.13.0",
-		"apollo-server-express": "^3.13.0",
+		"@graphql-tools/schema": "^10.0.2",
 		"bcryptjs": "^2.4.3",
-		"dotenv": "^16.4.5",
-		"graphql": "^16.8.1",
-		"graphql-upload": "^16.0.2",
+		"@typegoose/typegoose": "^12.1.0",
+		"apollo-server-express": "^3.12.0",
 		"jsonwebtoken": "^9.0.2",
+		"dotenv": "^16.3.1",
+		"express": "^4.18.2",
 		"mongodb": "^6.6.2",
-		"mongoose": "^8.1.3",
+		"mongoose": "^8.1.0",
 		"ncp": "^2.0.0",
+		"ts-node": "^10.9.2",
 		"python-shell": "^5.0.0",
-		"typescript": "^5.3.3"
+		"type-graphql": "^1.1.1"
 	},
 	"devDependencies": {
-		"@graphql-codegen/cli": "^5.0.2",
-		"@graphql-codegen/introspection": "^4.0.2",
-		"@graphql-codegen/typescript": "^4.0.4",
-		"@graphql-codegen/typescript-resolvers": "^4.0.4",
+		"@graphql-codegen/cli": "5.0.0",
+		"@graphql-codegen/introspection": "4.0.0",
+		"@graphql-codegen/typescript": "4.0.1",
+		"@graphql-codegen/typescript-resolvers": "4.0.1",
 		"@types/bcryptjs": "^2.4.6",
+		"@types/dotenv": "^8.2.0",
 		"@types/jsonwebtoken": "^9.0.6",
-		"@types/mongoose": "^5.11.97",
+		"@types/express": "^4.17.21",
 		"@types/multer": "^1.4.11",
-		"nodemon": "^3.0.3",
-		"serverless-offline": "^13.3.3",
-		"serverless-plugin-typescript": "^2.1.5",
-		"ts-node": "^10.9.2"
-	},
-	"author": "Brahm Van Houzen",
-	"license": "MIT"
+		"@types/mongoose": "^5.11.97",
+		"@types/node": "^20.10.5",
+		"graphql-scalars": "^1.22.4",
+		"ts-node": "^10.9.2",
+		"nodemon": "^2.0.3",
+		"typescript": "^5.2.2"
+	}
 }
