+++ schema/index.ts
@@ -1,0 +1,8 @@
+import Query from './queries';
+// import Mutation from './mutations';
+import { Resolvers } from '../__generated__/graphql';
+
+// const resolvers: Resolvers = { Query, Mutation };
+const resolvers: Resolvers = { Query };
+
+export default resolvers;
