+++ schema/mutations.ts
@@ -1,0 +1,8 @@
+// // import { MutationResolvers, MutationCreateEntryArgs, Entry } from '../__generated__/graphql';
+
+// const mutations = {
+	
+// };
+
+// export default mutations;
+ export {}
