--- src/connection/db.ts
@@ -1,15 +1,0 @@
-import { set, connect } from 'mongoose';
-import dotenv from 'dotenv';
-
-export const connectToDb = async () => {
-	dotenv.config();
-	const MONGODB_URI = process.env.MONGODB_URI ?? '';
-
-	try {
-		set('strictQuery', true);
-		await connect(MONGODB_URI).then(() => console.log('Connected to MongoDB'));
-	} catch (error) {
-		console.error('Error connecting to MongoDB', error);
-		throw new Error('Error connecting to MongoDB');
-	}
-};
