--- src/lib/Errors.ts
@@ -1,6 +1,0 @@
-class AuthenticationError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = 'AuthenticationError';
-  }
-}
