+++ tsconfig.json
@@ -1,14 +1,16 @@
 {
 	"compilerOptions": {
-		"target": "es5",
-		"module": "commonjs",
-		"allowJs": true,
+		/* Language and Environment */
+		"outDir": "./build",
+		"target": "es2016" /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */,
+		/* Modules */
+		"module": "commonjs" /* Specify what module code is generated. */,
+		"resolveJsonModule": true /* Enable importing .json files. */,
+		"esModuleInterop": true /* Emit additional JavaScript to ease support for importing CommonJS modules. This enables 'allowSyntheticDefaultImports' for type compatibility. */,
+		"forceConsistentCasingInFileNames": true /* Ensure that casing is correct in imports. */,
+
+		/* Type Checking */
 		"strict": true,
-		"esModuleInterop": true,
-		"skipLibCheck": true,
-		"forceConsistentCasingInFileNames": true,
-		"resolveJsonModule": true
-	},
-	"include": ["src/*.ts", "src/**/*.ts", "src/**/*.js"],
-	"exclude": ["node_modules", "src/generated/*.ts", "src/seed/*.ts", "src/**/*.d.ts"]
+		"skipLibCheck": true
+	}
 }
