+++ types.d.ts
@@ -1,0 +1,8 @@
+export interface IEntry {
+  date: Date;
+	financial: number;
+	fitness: number;
+	dietary: number;
+	social: number;
+	professional: number;
+}
