+++ utils/thumbtack_review_scraper/helper.ts
@@ -1,0 +1,9 @@
+export const extractThumbtackData = (reviewsArrArr: any) => {
+	console.log('extractThumbtackData reviewsArrArr');
+	let reviews: any = [];
+	reviewsArrArr.forEach((reviewsArr: any) => {
+		reviews.push(reviewsArr[0]);
+	});
+
+	return reviews;
+};
